requirements.txt
pandas==1.3.3
matplotlib==3.4.3
numpy==1.21.2
scikit-learn==0.24.2
seaborn==0.11.2
xgboost==1.4.2
optuna==2.10.0
scikit-optimize==0.8.1
scikit-optimize==0.8.1
PyYAML==5.4.1
